Cloud_Funtions_GCP/carga_datalake/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-bigquery
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2